# perf: Remove deep-copying the allocatable resource list